+++ packs/types.ts
@@ -32,3 +32,15 @@ export type SessionCustomer = {
   EmailRecuperacaoCompra: boolean;
   SourceCliente: string;
 };
+
+export interface Review {
+  ratingValue: number;
+  authorName: string;
+  reviewDescription: string;
+  authorCity: string;
+  productName: string;
+  productPhoto: string;
+  productLink: string;
+  additionalImage: string;
+  memberLevel?: string;
+}
